Mention the grate in the room's description, when visible.
At first glance, this would seem to go against my rule against adding
gameplay elements. But the code for showing the grate was still there,
it was just in an unused action routine. So I've changed that routine to
apply to PATH instead, because that's clearly what was intended.

PATH already had an action routine, but I call that from the new
PATH-FCN so that's still ok.

Just as a reminder, the grate is completely useless in this version
since you can never find the key for it.