Fixed fighting with the thief.
This bug was caused by a faulty end-of-table check for the VILLAINS
table. An LTABLE is a 1-based list with the number of elements stored in
index 0. Therefore, only exit the loop after going *past* LEN.

This bug seems to be present in the full game as well, but in that case
there *is* a third entry in the table for the cyclopes. Which is
interesting, because you were able to fight the cyclops in the MDL
version of Zork, and Zork I does contain the messages for fighting the
cyclops.

(Though since the cyclops is unarmed, the game would consider him
helpless and unable to defend himself.)